refactor code according to origin on github
