+++ lib/tools/games/game_of_life/logic/generate_rle/font_data.dart
@@ -1,0 +1,123 @@
+// https://github.com/tlrobinson/life-gen/blob/master/font.txt
+//
+// # Copyright (c) 2009 Thomas Robinson <tlrobinson.net>
+// #
+// # Permission is hereby granted, free of charge, to any person
+// # obtaining a copy of this software and associated documentation
+// # files (the "Software"), to deal in the Software without
+// # restriction, including without limitation the rights to use,
+// # copy, modify, merge, publish, distribute, sublicense, and/or sell
+// # copies of the Software, and to permit persons to whom the
+// # Software is furnished to do so, subject to the following
+// # conditions:
+// #
+// # The above copyright notice and this permission notice shall be
+// # included in all copies or substantial portions of the Software.
+// #
+// # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// # EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// # OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// # NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// # HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+// # WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// # FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+// # OTHER DEALINGS IN THE SOFTWARE.
+
+part of 'package:gc_wizard/tools/games/game_of_life/logic/game_of_life_rle_generate.dart';
+
+const _RLE_FONT_DATA =
+    '0000000000000000000000000000000000000000,'
+    '0000800080008000800080000000800000000000,'
+    '0000a000a0000000000000000000000000000000,'
+    '000000005000f80050005000f800500000000000,'
+    '20007000a800a00070002800a800700020000000,'
+    '04004400a8004800120015002200200000000000,'
+    '00006000900090006000a8009000ec0000000000,'
+    '0000800080000000000000000000000000000000,'
+    '0000400080008000800080008000400000000000,'
+    '0000800040004000400040004000800000000000,'
+    '000088005000f800500088000000000000000000,'
+    '0000000020002000f80020002000000000000000,'
+    '0000000000000000000000000000400080000000,'
+    '0000000000000000f00000000000000000000000,'
+    '0000000000000000000000000000800000000000,'
+    '0000100010002000200040004000800080000000,'
+    '0000700088008800880088008800700000000000,'
+    '00004000c0004000400040004000e00000000000,'
+    '0000700088000800300040008000f80000000000,'
+    '0000700088000800300008008800700000000000,'
+    '00001000300050009000f8001000100000000000,'
+    '0000f00080008000e00010001000e00000000000,'
+    '0000700088008000f00088008800700000000000,'
+    '0000f80008001000200020004000400000000000,'
+    '0000700088008800700088008800700000000000,'
+    '0000700088008800780008001000600000000000,'
+    '0000000080000000000000008000000000000000,'
+    '0000000040000000000000004000800000000000,'
+    '0000100020004000800040002000100000000000,'
+    '000000000000f8000000f8000000000000000000,'
+    '0000800040002000100020004000800000000000,'
+    '0000700088001000200020000000200000000000,'
+    '000000007c0082009a00aa00ba004e0000000000,'
+    '0000700088008800f80088008800880000000000,'
+    '0000f00088008800f00088008800f00000000000,'
+    '0000700088008000800080008800700000000000,'
+    '0000f00088008800880088008800f00000000000,'
+    '0000f80080008000f00080008000f80000000000,'
+    '0000f80080008000f00080008000800000000000,'
+    '0000700088008000980088008800700000000000,'
+    '0000880088008800f80088008800880000000000,'
+    '0000f80020002000200020002000f80000000000,'
+    '0000080008000800080088008800700000000000,'
+    '000088009000a000c000a0009000880000000000,'
+    '0000800080008000800080008000f80000000000,'
+    '00008800d800a800880088008800880000000000,'
+    '000088008800c800a80098008800880000000000,'
+    '0000700088008800880088008800700000000000,'
+    '0000f00088008800f00080008000800000000000,'
+    '00007800840084008400840078000c0000000000,'
+    '0000f00088008800f000a0009000880000000000,'
+    '0000700088008000700008008800700000000000,'
+    '0000f80020002000200020002000200000000000,'
+    '0000880088008800880088008800700000000000,'
+    '0000880088008800880088005000200000000000,'
+    '000088008800a800a800a8005000500000000000,'
+    '0000880088005000200050008800880000000000,'
+    '0000880088005000200020002000200000000000,'
+    '0000f80008001000200040008000f80000000000,'
+    '0000c00080008000800080008000c00000000000,'
+    '0000800080004000400020002000100010000000,'
+    '0000c00040004000400040004000c00000000000,'
+    '4000a000a0000000000000000000000000000000,'
+    '0000000000000000000000000000f80000000000,'
+    '0000800040000000000000000000000000000000,'
+    '0000000000007000900090009000780000000000,'
+    '000080008000e000900090009000e00000000000,'
+    '0000000000007000800080008000700000000000,'
+    '0000100010007000900090009000700000000000,'
+    '00000000000070008800f8008000780000000000,'
+    '300040004000f000400040004000400080000000,'
+    '0000000000007000900090009000700090006000,'
+    '000080008000e000900090009000900000000000,'
+    '000040000000c000400040004000e00000000000,'
+    '0000100000003000100010001000100090006000,'
+    '0000800080009000a000e0009000900000000000,'
+    '0000800080008000800080008000800000000000,'
+    '000000000000d000a800a800a800a80000000000,'
+    '000000000000e000900090009000900000000000,'
+    '0000000000007000880088008800700000000000,'
+    '000000000000e000900090009000e00080008000,'
+    '0000000000006000900090009000700010001000,'
+    '000000000000b000c00080008000800000000000,'
+    '0000000000007000800060001000e00000000000,'
+    '000040004000f000400040004000300000000000,'
+    '0000000000009000900090009000700000000000,'
+    '0000000000008800880050005000200000000000,'
+    '0000000000008800a800a8005000500000000000,'
+    '0000000000008800500020005000880000000000,'
+    '0000000000009000900090009000700010006000,'
+    '000000000000f800100020004000f80000000000,'
+    '00000000600040004000c0004000400060000000,'
+    '0000800080008000800080008000800000000000,'
+    '00000000c00040004000600040004000c0000000,'
+    '00004000a8001000000000000000000000000000';
